view: wait until surface is mapped to send enter
Based on observation, this is what weston does. I'm not sure if it's
useful to send enter on surface creation before the surface is mapped,
howver the wayland protocol seems to encourage it, saying explicitly
enter should be send on "creation".

This commit is needed for firefox to render popups when using
multiple monitors and different output scales.